MINOR: Check required user/group before starting zrancid-ttyd service
We let ttyd drop privileges using -u/-g options so it can create a
socket writable by the apache group, so systemd will not check itself
if the zrancid-ttyd user exist before starting the process. Also we
noticed that if the apache group does not exist, ttyd does an infinite
error log loop. The patch adds ExecStartPre commands to check if
required user/group exist. It is better especially because we do not
require httpd in the package (to relax dependencies).