MEDIUM: Override reconnect overlay style in sample ttyd configuration
This ugly hack uses apache mod_substitute to modify the CSS embedded
in ttyd's index.html in order to display a more appropriate font in
the overlay panel.
That way we can stop maintaining our own ttyd RPM, because it is now
present in EPEL repos. Our ttyd RPM used to provided an alternate
index.html with CSS style override for the overlay panel.